b=15670 i=adilger i=johann
land attachment 16523

fixes this compile warning (which is treated as an error -Werror on XT3's):
during lustrefs-1.6-ss builds:

lustrefs-1.6-ss/lustre/mds/mds_lov.c: In function `mds_lov_read_objids':
mds_lov.c:198: warning: int format, different type arg (arg 10)

+++ lustre/mds/mds_lov.c
@@ -194,7 +194,7 @@ static int mds_lov_read_objids(struct obd_device *obd)
                 RETURN(0);
 
         page = (size/(OBJID_PER_PAGE()*sizeof(obd_id)))+1;
-        CDEBUG(D_INFO, "file size %d pages %d\n", size, page);
+        CDEBUG(D_INFO, "file size %d pages %d\n", (int)size, page);
         for(i=0; i < page; i++) {
                 obd_id *data =  mds->mds_lov_page_array[i];
                 loff_t off_old = off;
